mox: added python3 specific way to detect if symbol is a method
In python 3 there is no way differentiate a function from a method, in
python 2 there were unbound and bound methods, now only functions and bound
methods exist.

This patch in python 3 to detect if a symbol is a method checks if the first
argument's name is 'self', so if you have a function with argument self then
this isn't gonna work for you, in any case having a function with a 'self'
argument is ugly, so should probably use 'self_'.